requirements.txt
APScheduler==3.6.3
asn1crypto==1.3.0
beautifulsoup4==4.8.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
Click==7.0
configreader==0.0.6
cryptography==2.8
cycler==0.10.0
dnspython==1.16.0
Flask==1.1.1
Flask-Cors==3.0.8
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
psutil==5.7.0
pycparser==2.20
pymongo==3.9.0
pyOpenSSL==19.1.0
pyparsing==2.4.6
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
six==1.14.0
soupsieve==2.0
SQLAlchemy==1.3.13
tzlocal==2.0.0
urllib3==1.25.8
Werkzeug==1.0.1
win-inet-pton==1.1.0
wincertstore==0.2
#winkerberos==0.7.0
aanalytics2==0.3.0.post2
argcomplete==1.12.3
asgiref==3.5.0
attrs==21.4.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bidict==0.21.4
blinker==1.4
boto3==1.21.13
botocore==1.24.13
bs4==0.0.1
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorlog==6.6.0
cryptography==36.0.1
dicttoxml==1.7.4
distlib==0.3.4
Django==4.0.3
dnspython==2.2.1
docopt==0.6.2
et-xmlfile==1.1.0
filelock==3.6.0
Flask==2.0.3
Flask-Cors==3.0.10
Flask-Mail==0.9.1
Flask-SocketIO==5.1.1
google-ads==15.0.0
google-api-core==2.7.1
google-auth==1.35.0
google-auth-oauthlib==0.5.0
google-trans-new==1.1.9
googleads==31.0.0
googleapis-common-protos==1.55.0
grpcio==1.44.0
httplib2==0.20.4
idna==3.3
iniconfig==1.1.1
isodate==0.6.1
itsdangerous==2.1.0
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
lxml==4.8.0
MarkupSafe==2.1.0
mypy-boto3-chime==1.21.0
nltk==3.7
nox==2022.1.7
npm==0.1.1
numpy==1.22.2
numpyencoder==0.3.0
oauth2client==4.1.3
oauthlib==3.2.0
openpyxl==3.0.9
optional-django==0.1.0
packaging==21.3
pandas==1.4.1
passlib==1.7.4
pathlib==1.0.1
pathlib2==2.3.7.post1
paypalrestsdk==1.13.1
pipenv==2022.3.28
pipreqs==0.4.11
platformdirs==2.5.1
plotly==5.8.2
pluggy==1.0.0
proto-plus==1.18.1
protobuf==3.17.3
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycountry==22.3.5
pycparser==2.21
PyJWT==2.3.0
pymongo==3.12.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
pytest==7.0.1
python-dateutil==2.8.2
python-engineio==4.3.1
python-geoip==1.2
python-geoip-geolite2==2015.303
python-socketio==5.5.2
pytz==2021.3
PyYAML==5.4.1
rake-nltk==1.0.6
regex==2022.3.2
requests==2.27.1
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rsa==4.8
s3transfer==0.5.2
scikit-learn==1.0.2
scipy==1.8.0
six==1.15.0
sklearn==0.0
sort-dataframeby-monthorweek==0.4
sorted-months-weekdays==0.2
soupsieve==2.3.2.post1
sqlparse==0.4.2
tenacity==8.0.1
textblob==0.17.1
threadpoolctl==3.1.0
tomli==2.0.1
tqdm==4.63.0
typing-extensions==4.1.1
urllib3==1.26.8
virtualenv==20.13.3
virtualenv-clone==0.5.7
Werkzeug==2.0.3
xmltodict==0.12.0
yarg==0.1.9
zeep==4.1.0
